Cleanup bitsim, document hypo.
